SW compiler warnings fix
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@3646 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Algorithm/src/smith_waterman/SmithWatermanReportCallback.cpp
@@ -388,7 +388,7 @@ void SmithWatermanReportCallbackMAImpl::alignSequences(QList<U2MsaGap>& refSeque
     bool lastSymbolIsGapPtrn = false;
     quint32 intervalStart;
     quint32 intervalEnd;
-    for (quint32 i = 0; i < pairwiseAlignment.length(); ++i) {
+    for (quint32 i = 0; i < static_cast<quint32>(pairwiseAlignment.length()); ++i) {
         switch (pairwiseAlignment[i]) {
         case SmithWatermanResult::DIAG:
             if (lastSymbolIsGapRef == true) {
